requirements.txt
streamlit
pandas
python-dotenv
chromadb
openai